update: refactor some code & add post service

+++ api-persistence/src/main/java/com/api/persistence/repository/system/SysUserPostRepository.java
@@ -32,7 +32,7 @@ public interface SysUserPostRepository
    * @param postId post id
    * @return count
    */
-  long countByPost_PostId(Long postId);
+  int countByPost_PostId(Long postId);
 
   /**
    * Delete user-post associations by a list of userIds.
